fix(pr): address follow-up review comments
Preserve non-CCC hooks while refreshing local hook installs.

Keep legacy sessions on Claude defaults instead of active Codex providers.

Use portable file locks for hook and ledger state, log Telegram callback edit failures, and widen message IDs.

+++ pkg/listen/new.go
@@ -275,7 +275,7 @@ func newAgentButtons(cfg *configpkg.Config, sessionName string) [][]telegram.Inl
 	return buttons
 }
 
-func sendNewProviderSelection(cfg *configpkg.Config, chatID, threadID int64, messageID int, sessionName, agentName string) {
+func sendNewProviderSelection(cfg *configpkg.Config, chatID, threadID int64, messageID int64, sessionName, agentName string) {
 	buttons := newProviderButtonsForAgent(cfg, sessionName, agentName)
 	if len(buttons) == 0 {
 		msg := fmt.Sprintf("❌ No %s provider/model is configured.", agentOptionLabel(cfg, agentName))
